pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coacoa"

version = "0.1.0"

description = "Context-Optimised Agentic Coding Assistant"

authors = [
  { name = "YOUR NAME", email = "your@email.com" }
]

license = { text = "Apache-2.0" }

requires-python = ">=3.9"

dependencies = [
  "typer>=0.9.0",
  "importlib_resources>=6.0 ; python_version < '3.11'",
  "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.21.0",
    "coverage[toml]>=7.0.0",
    "pre-commit>=3.0.0",
    "tox>=4.0.0",
    "typer>=0.9.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.21.0",
    "coverage[toml]>=7.0.0",
]

[project.scripts]
coacoa = "coacoa.__main__:app"

[tool.setuptools.package-data]
"coacoa.scaffold" = [".coacoa/**/*"]